requirements.txt
# requirements.txt

flask==0.12.2
flask_basicauth==0.2.0
python-dateutil==2.6.1
nordpool==0.2
sqlalchemy==1.2.2
flask_wtf==2.1